# openapi.yaml
openapi: 3.0.3
info:
  title: Task Management API
  description: RESTful API for managing tasks and user assignments
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/tasks:
    get:
      tags:
        - Tasks
      summary: List tasks
      description: Get all tasks in organizations the user is a member of
      operationId: listTasks
      security:
        - cookieAuth: []
      parameters:
        - name: organization_id
          in: query
          description: Filter tasks by organization ID (optional)
          schema:
            type: integer
            format: int64
        - name: assigned_to_me
          in: query
          description: Filter tasks assigned to the current user
          schema:
            type: boolean
            default: false
        - name: due_today
          in: query
          description: Filter tasks due today
          schema:
            type: boolean
            default: false
        - name: status
          in: query
          description: Filter tasks by status (TODO or DONE)
          schema:
            type: string
            enum: [TODO, DONE]
        - name: sort
          in: query
          description: Sort order. Use 'due_date' to sort by due date (nearest first, NULL last). Default is by created_at descending.
          schema:
            type: string
            enum: [due_date]
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags:
        - Tasks
      summary: Create task
      description: Create a new task in an organization
      operationId: createTask
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - organization_id
              properties:
                title:
                  type: string
                  example: Complete project documentation
                description:
                  type: string
                  example: Write comprehensive API documentation
                due_date:
                  type: string
                  format: date-time
                  example: 2025-12-31T23:59:59Z
                organization_id:
                  type: integer
                  format: int64
                  example: 1
      responses:
        "201":
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: You are not a member of this organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/tasks/generate:
    post:
      tags:
        - Tasks
      summary: Generate task suggestions using AI
      description: Generate task suggestions from text using OpenAI
      operationId: generateTasks
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
                - organization_id
              properties:
                text:
                  type: string
                  example: "プロジェクトの要件定義を作成し、チームメンバーに共有する。来週までにデザインモックアップを完成させる。"
                organization_id:
                  type: integer
                  format: int64
                  example: 1
      responses:
        "200":
          description: Tasks generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: You are not a member of this organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: AI service is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/tasks/{id}:
    get:
      tags:
        - Tasks
      summary: Get task
      description: Get a specific task by ID
      operationId: getTask
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      tags:
        - Tasks
      summary: Update task (partial update)
      description: |
        Update task fields. Only fields included in the request body will be updated.
        Omitted fields will keep their current values.
        To clear due_date, explicitly send null.
      operationId: updateTask
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Updated task title
                description:
                  type: string
                  example: Updated description
                due_date:
                  type: string
                  format: date-time
                  nullable: true
                  example: 2025-12-31T23:59:59Z
            examples:
              updateTitle:
                summary: Update only title
                value:
                  title: "New Title"
              updateMultiple:
                summary: Update multiple fields
                value:
                  title: "New Title"
                  description: "New Description"
                  due_date: "2025-12-31T23:59:59Z"
              clearDueDate:
                summary: Clear due date
                value:
                  due_date: null
      responses:
        "200":
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags:
        - Tasks
      summary: Delete task
      description: Delete a task (only creator can delete)
      operationId: deleteTask
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Task deleted successfully
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Only creator can delete task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/tasks/{id}/assign:
    post:
      tags:
        - Tasks
      summary: Assign users to task
      description: Assign one or more users to a task (only creator can assign)
      operationId: assignTask
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_ids
              properties:
                user_ids:
                  type: array
                  items:
                    type: integer
                    format: int64
                  example: [2, 3, 4]
      responses:
        "200":
          description: Users assigned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Users assigned successfully
                  assignments:
                    type: array
                    items:
                      $ref: "#/components/schemas/TaskAssignment"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Only creator can assign users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Task not found or user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/tasks/{id}/unassign:
    post:
      tags:
        - Tasks
      summary: Unassign users from task
      description: Remove user assignments from a task (only creator can unassign)
      operationId: unassignTask
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_ids
              properties:
                user_ids:
                  type: array
                  items:
                    type: integer
                    format: int64
                  example: [2, 3]
      responses:
        "200":
          description: Users unassigned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Users unassigned successfully
                  assignments:
                    type: array
                    items:
                      $ref: "#/components/schemas/TaskAssignment"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Only creator can unassign users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/tasks/{id}/toggle-status:
    post:
      tags:
        - Tasks
      summary: Toggle task status
      description: Toggle task status between TODO and DONE (only creator or assigned users can toggle)
      operationId: toggleTaskStatus
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Task status toggled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Task status updated successfully
                  status:
                    type: string
                    enum: [TODO, DONE]
                    example: DONE
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Only creator or assigned users can toggle status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    User:
      type: object
      required:
        - id
        - username
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: johndoe
        created_at:
          type: string
          format: date-time
          example: 2025-01-01T00:00:00Z
        updated_at:
          type: string
          format: date-time
          example: 2025-01-01T00:00:00Z
    Organization:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: My Team
        invite_code:
          type: string
          example: "5dc6-6411-e229"
...
    Task:
      type: object
      required:
        - id
        - title
        - description
        - status
        - creator_id
        - organization_id
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: Complete project documentation
        description:
          type: string
          example: Write comprehensive API documentation
        status:
          type: string
          enum: [TODO, DONE]
          example: TODO
          description: Task completion status
        due_date:
          type: string
          format: date-time
          nullable: true
          description: Optional due date for the task. Can be null.
          example: 2025-12-31T23:59:59Z
        creator_id:
          type: integer
          format: int64
          example: 1
        organization_id:
          type: integer
          format: int64
          example: 1
        created_at:
          type: string
          format: date-time
          example: 2025-01-01T00:00:00Z
        updated_at:
          type: string
          format: date-time
          example: 2025-01-01T00:00:00Z
        creator:
          description: Task creator user object (only included when explicitly loaded)
          $ref: "#/components/schemas/User"
        organization:
          description: Organization object (only included when explicitly loaded)
          $ref: "#/components/schemas/Organization"
        assignments:
          type: array
          description: List of task assignments (only included when explicitly loaded)
          items:
            $ref: "#/components/schemas/TaskAssignment"
    TaskAssignment:
      type: object
      required:
        - user
      properties:
        user:
          $ref: "#/components/schemas/User"
    TaskListItem:
      type: object
      required:
        - id
        - title
        - description
        - status
        - creator_id
        - created_at
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: Complete project documentation
        description:
          type: string
          example: Write comprehensive API documentation
        status:
          type: string
          enum: [TODO, DONE]
          example: TODO
        due_date:
          type: string
          format: date-time
          nullable: true
          example: 2025-12-31T23:59:59Z
        creator_id:
          type: integer
          format: int64
          example: 1
        creator:
          $ref: "#/components/schemas/User"
        created_at:
          type: string
          format: date-time
          example: 2025-01-01T00:00:00Z
    TaskListResponse:
      type: object
      required:
        - tasks
        - page
        - page_size
        - total_count
        - total_pages
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/TaskListItem"
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total_count:
          type: integer
          format: int64
          example: 42
        total_pages:
          type: integer
          example: 3
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Error message describing what went wrong